requirements.txt
black
ebaysdk
discord.py
python-dotenv
spacy
pytest
bs4
price-parser
currencyconverter